+++ index.js
@@ -5,21 +5,34 @@ var fs = require('fs')
 var path = require('path')
 var spawn = require('child_process').spawn
 var RSVP = require('rsvp')
+var request = require('sync-request')
 
-function runCommand (/* child_process.exec args */) {
+function runCommand (command, args, ignoreStdError) {
   return new RSVP.Promise(function (resolve, reject) {
-    var child = spawn('ember', ['test'])
+    var child = spawn(command, args)
     child.stdout.on('data', function (data) {
       console.log(data.toString())
     })
     child.stderr.on('data', function (data) {
-      reject(data.toString())
+      if (ignoreStdError || data.toString().indexOf('fs: re-evaluating native module sources is not supported.') > -1) {
+        // Use ignoreStdError only to get around this issue https://github.com/ciena-blueplanet/ember-cli-visual-acceptance/issues/25
+        console.log(data.toString())
+      } else {
+        reject(data.toString())
+      }
     })
 
     child.on('exit', function (code) {
       resolve()
     })
   })
+}
+
+function base64Encode (file) {
+  // read binary data
+  var bitmap = fs.readFileSync(file)
+  // convert binary data to base64 encoded string
+  return new Buffer(bitmap).toString('base64')
 }
 
 function compareVersions (installed, required) {
@@ -58,6 +71,19 @@ function mkdirpSync (dirpath) {
   for (let i = 1; i <= parts.length; i++) {
     mkdirSync(path.join.apply(null, parts.slice(0, i)))
   }
+}
+
+function appendToReport (req, res, options) {
+  try {
+    if (process.env.REPORT_PATH) {
+      var report = fs.readFileSync(process.env.REPORT_PATH)
+      report = report.toString().replace(/(<\/div>\s<\/body>\s<\/HTML>)/i, req.body.report + '$1')
+      fs.writeFileSync(process.env.REPORT_PATH, report)
+    }
+  } catch (e) {
+    console.log(e)
+  }
+  res.send()
 }
 
 function isTargetBrowser (req, res, targetBrowsers) {
@@ -115,33 +141,122 @@ function getImage (req, res, options) {
   }
 }
 
+function buildReport (params) {
+  return runCommand('phantomjs', ['vendor/html-to-image.js', 'visual-acceptance-report/report.html']).then(function (params) {
+    console.log('Sending to github')
+    var image = base64Encode('images/output.png').replace('data:image\/\w+;base64,', '')
+    function uploadToImgur (image, reportPath) {
+      var imgurAlbum = []
+      var result = []
+      var imgurClientID = 'e39f00905b80937'
+      var imgurApiOptions = {
+        'headers': {
+          'Content-Type': 'application/json',
+          'Authorization': 'Client-ID ' + imgurClientID
+        },
+        'json': {
+          'type': 'base64',
+          'image': image.replace('data:image\/\w+;base64,', '')
+        }
+
+      }
+      var response = request('POST', 'https://api.imgur.com/3/image', imgurApiOptions)
+      var reportID = JSON.parse(response.getBody())
+      imgurAlbum.push(reportID.data.id)
+      result.push(reportID.data.link)
+      var report = fs.readFileSync(reportPath).toString()
+      var regexImageData = /src=\"data\:image\/png;base64\,[^"]*"/ig
+      var matches = report.match(regexImageData)
+      for (var i = 0; i < matches.length; i++) {
+        imgurApiOptions = {
+          'headers': {
+            'Content-Type': 'application/json',
+            'Authorization': 'Client-ID ' + imgurClientID
+          },
+          'json': {
+            'type': 'base64',
+            'image': matches[i].replace('src="data:image/png;base64', '').replace('"', '')
+          }
+
+        }
+        response = request('POST', 'https://api.imgur.com/3/image', imgurApiOptions)
+        imgurAlbum.push(JSON.parse(response.getBody()).data.id)
+      }
+      // create album
+      imgurApiOptions = {
+        'headers': {
+          'Content-Type': 'application/json',
+          'Authorization': 'Client-ID ' + imgurClientID
+        },
+        'json': {
+          'ids': imgurAlbum
+        }
+
+      }
+      response = request('POST', 'https://api.imgur.com/3/album', imgurApiOptions)
+      result.push('http://imgur.com/a/' + JSON.parse(response.getBody()).data.id)
+      return result
+    }
+
+    var imgurResponse = uploadToImgur(image, 'visual-acceptance-report/report.html')
+    var githubApiPostOptions = {
+      'headers': {
+        'user-agent': 'visual-acceptance',
+        'Authorization': 'token ' + process.env.VISUAL_ACCEPTANCE_TOKEN
+      },
+      'json': {
+        'body': '![PR ember-cli-visual-acceptance Report](' + imgurResponse[0] + ') \n [Imgur Album](' + imgurResponse[1] + ')'
+      }
+    }
+
+    var githubApiGetOptions = {
+      'headers': {
+        'user-agent': 'visual-acceptance',
+        'Authorization': 'token ' + process.env.VISUAL_ACCEPTANCE_TOKEN
+      }
+    }
+    var url = 'https://api.github.com/repos/' + process.env.TRAVIS_REPO_SLUG + '/issues/' + process.env.TRAVIS_PULL_REQUEST + '/comments'
+    var response = request('GET', url, githubApiGetOptions)
+    var bodyJSON = JSON.parse(response.getBody().toString())
+    for (var i = 0; i < bodyJSON.length; i++) {
+      if (bodyJSON[i].body.indexOf('![PR ember-cli-visual-acceptance Report]') > -1) {
+        url = bodyJSON[i].url
+        break
+      }
+    }
+    response = request('POST', url, githubApiPostOptions)
+    return response
+  })
+}
+
 module.exports = {
   name: 'ember-cli-visual-acceptance',
   included: function (app) {
     this._super.included(app)
-    if (process.env.EMBER_CLI_FASTBOOT !== 'true') {
+    if (app) {
       app.import(app.bowerDirectory + '/resemblejs/resemble.js', {
         type: 'test'
       })
-      app.import(app.bowerDirectory + '/detectjs/src/detect.js', {
+      app.import(path.join('vendor', 'bluebird', 'js', 'browser', 'bluebird.min.js'), {
         type: 'test'
       })
-      app.import('vendor/html2canvas.js', {
+      app.import(path.join('vendor', 'jquery.min.js'), {
         type: 'test'
       })
-      app.import('vendor/VisualAcceptance.js', {
+      app.import(path.join('vendor', 'html2canvas.js'), {
+        type: 'test'
+      })
+      app.import(path.join('vendor', 'VisualAcceptance.js'), {
+        type: 'test'
+      })
+      app.import(path.join('vendor', 'visual-acceptance-report.css'), {
+        type: 'test'
+      })
+      app.import(path.join('vendor', 'detect.js'), {
         type: 'test'
       })
     }
-    app.import('vendor/dist/css/materialize.min.css', {
-      type: 'test'
-    })
-    app.import('vendor/dist/js/materialize.min.js', {
-      type: 'test'
-    })
-    app.import('vendor/visual-acceptance-report.css', {
-      type: 'test'
-    })
+
     if (app.options.visualAcceptanceOptions) {
       this.imageDirectory = app.options.visualAcceptanceOptions.imageDirectory || 'visual-acceptance'
       this.targetBrowsers = app.options.visualAcceptanceOptions.targetBrowsers || []
@@ -151,15 +266,16 @@ module.exports = {
   targetBrowsers: [],
   middleware: function (app, options) {
     app.use(bodyParser.urlencoded({
-      extended: true
+      limit: '50mb',
+      extended: true,
+      parameterLimit: 50000
     }))
-    app.use(bodyParser.json())
+    app.use(bodyParser.json({
+      limit: '50mb'
+    }))
+
     app.get('/image', function (req, res) {
       getImage(req, res, options)
-    })
-
-    app.get('/istargetbrowser', function (req, res) {
-      isTargetBrowser(req, res, options.targetBrowsers)
     })
 
     app.post('/image', function (req, res) {
@@ -171,6 +287,16 @@ module.exports = {
     })
     app.post('/fail', function (req, res) {
       misMatchImage(req, res, options)
+    })
+    app.post('/report', function (req, res) {
+      appendToReport(req, res, options)
+    })
+    app.get('/istargetbrowser', function (req, res) {
+      isTargetBrowser(req, res, options.targetBrowsers)
+    })
+
+    app.get('/visual-acceptance', function (req, res) {
+      res.sendfile('visual-acceptance-report/report.html')
     })
   },
   testemMiddleware: function (app) {
@@ -194,9 +320,94 @@ module.exports = {
 
   includedCommands: function () {
     return {
+      'build-report': {
+        name: 'build-report',
+        aliases: ['br'],
+        description: 'Create report',
+        works: 'insideProject',
+        availableOptions: [{
+          name: 'report-directory',
+          type: String,
+          default: 'visual-acceptance-report',
+          description: 'Create Report off visual acceptance tests'
+        }],
+        run: function (options, rawArgs) {
+          var root = this.project.root
+
+          function deleteFolderRecursive (path) {
+            if (fs.existsSync(path)) {
+              fs.readdirSync(path).forEach(function (file, index) {
+                var curPath = path + '/' + file
+                if (fs.lstatSync(curPath).isDirectory()) { // recurse
+                  deleteFolderRecursive(curPath)
+                } else { // delete file
+                  fs.unlinkSync(curPath)
+                }
+              })
+              fs.rmdirSync(path)
+            }
+          }
+
+          deleteFolderRecursive(path.join(root, options.reportDirectory))
+          mkdirpSync(options.reportDirectory)
+          var reportPath = options.reportDirectory + '/' + 'report.html'
+          fs.writeFileSync(reportPath, `<HTML>
+<HEAD>
+<TITLE>Visual Acceptance report </TITLE>
+</HEAD>
+<BODY>
+<style>
+  .visual-acceptance.title  {
+    padding-top: 30px;
+    font-size: 29px;
+    color: #404041;
+    padding-left: 30px;
+  }
+  .visual-acceptance-container {
+    padding-left: 30px;
+  }
+  .visual-acceptance-container > .test {
+    padding-left: 20px;
+  }
+  .visual-acceptance-container > .test > .list-name {
+    font-size: 18px;
+    color: #33424F;
+    padding-top: 20px;
+  }
+
+  .visual-acceptance-container > .test > .list-subtitle {
+    float: left;
+  }
+  .visual-acceptance-container > .test > .additional-info {
+    font-size: 14px;
+    color: #929497;
+    padding-bottom: 20px;
+  }
+
+  .images .image{
+      float: left;
+      text-decoration:none;
+  }
+  img {
+    width: 160px;
+    height: 160px;
+    padding-right: 10px;
+  }
+</style>
+  <div class="visual-acceptance title"> Visual Acceptance tests: </div>
+  <div class="visual-acceptance-container"> 
+</div>
+</BODY>
+</HTML>`)
+
+          process.env.PR_API = options.prApiUrl
+          process.env.REPORT_PATH = reportPath
+          return runCommand('ember', ['test'])
+        }
+      },
       'new-baseline': {
         name: 'new-baseline',
-        aliases: ['new-baseline'],
+        aliases: ['nb'],
         description: 'Create new baseline',
         works: 'insideProject',
         availableOptions: [{
@@ -207,7 +418,7 @@ module.exports = {
         }],
         run: function (options, rawArgs) {
           var root = this.project.root
-          var execOptions = { cwd: root }
+
           function deleteFolderRecursive (path) {
             if (fs.existsSync(path)) {
               fs.readdirSync(path).forEach(function (file, index) {
@@ -223,7 +434,75 @@ module.exports = {
           }
 
           deleteFolderRecursive(path.join(root, options.imageDirectory))
-          return runCommand('ember test', execOptions)
+          return runCommand('ember', ['test'])
+        }
+      },
+      'travis-visual-acceptance': {
+        name: 'travis-visual-acceptance',
+        aliases: ['tva'],
+        description: 'Run visual-acceptance based off Travis message',
+        works: 'insideProject',
+        availableOptions: [{
+          name: 'image-directory',
+          type: String,
+          default: 'visual-acceptance',
+          description: 'The ember-cli-visual-acceptance directory where images are save'
+        }, {
+          name: 'branch',
+          type: String,
+          default: 'master',
+          description: 'branch to push to'
+        }],
+        run: function (options, rawArgs) {
+          let requestOptions = {
+            'headers': {
+              'user-agent': 'ciena-frost',
+              'Authorization': 'token ' + process.env.RO_GH_TOKEN
+            }
+          }
+
+          if (!process.env.RO_GH_TOKEN || !process.env.TRAVIS_REPO_SLUG) {
+            console.log('No github token found or Travis found. Just running ember test')
+            return runCommand('ember', ['test'])
+          }
+          var repoSlug = process.env.TRAVIS_REPO_SLUG
+
+          var prNumber = process.env.TRAVIS_PULL_REQUEST
+          var url = 'https://api.github.com/repos/' + repoSlug + '/pulls/' + prNumber
+          var res = request('GET', url, requestOptions)
+          var travisMessage = res.body
+          if (/\#new\-baseline\#/.exec(travisMessage)) {
+            console.log('Creating new baseline')
+            return runCommand('ember', ['new-baseline', '--image-directory=' + options.imageDirectory]).then(function (params) {
+              if (prNumber === false) {
+                console.log('Git add')
+                return runCommand('git', ['add', options.imageDirectory + '/*']).then(function (params) {
+                  console.log('Git commit')
+                  return runCommand('git', ['commit', '-m', '"Adding new baseline images [ci skip]"']).then(function (params) {
+                    console.log('Git push')
+                    return runCommand('git', ['push', 'origin', 'HEAD:' + options.branch], true)
+                  })
+                })
+              }
+            })
+          } else if (prNumber !== false && prNumber !== 'false' && process.env.VISUAL_ACCEPTANCE_TOKEN) {
+            return runCommand('ember', ['br']).then(buildReport, function (params) {
+              return buildReport(params).then(function (params) {
+                throw new Error('Exit 1')
+              })
+            })
+          } else if ((prNumber === false || prNumber === 'false')) {
+            return runCommand('ember', ['test']).then(function (params) {
+              console.log('Git add')
+              return runCommand('git', ['add', options.imageDirectory + '/*']).then(function (params) {
+                console.log('Git commit')
+                return runCommand('git', ['commit', '-m', '"Adding new baseline images [ci skip]"']).then(function (params) {
+                  console.log('Git push')
+                  return runCommand('git', ['push', 'origin', 'HEAD:' + options.branch], true)
+                })
+              })
+            })
+          }
         }
       }
     }