+++ CHANGELOG.md
@@ -1,3 +1,44 @@
+# 0.1.9
+No CHANGELOG section found in Pull Request description.
+Use a `# CHANGELOG` section in your Pull Request description to auto-populate the `CHANGELOG.md`
+
+# 0.1.8
+No CHANGELOG section found in Pull Request description.
+Use a `# CHANGELOG` section in your Pull Request description to auto-populate the `CHANGELOG.md`
+
+# 0.1.7
+No CHANGELOG section found in Pull Request description.
+Use a `# CHANGELOG` section in your Pull Request description to auto-populate the `CHANGELOG.md`
+
+# 0.1.6
+No CHANGELOG section found in Pull Request description.
+Use a `# CHANGELOG` section in your Pull Request description to auto-populate the `CHANGELOG.md`
+
+# 0.1.5
+* Remove prApiUrl. Instead comment with an account that has no public repo access. So scope can only comment
+* serve html report on `/visual-acceptance` fixes https://github.com/ciena-blueplanet/ember-cli-visual-acceptance/issues/7
+
+# 0.1.4
+No CHANGELOG section found in Pull Request description.
+Use a `# CHANGELOG` section in your Pull Request description to auto-populate the `CHANGELOG.md`
+
+# 0.1.3
+* Temporarily resolve this issue https://github.com/ciena-blueplanet/ember-cli-visual-acceptance/issues/27
+
+
+# 0.1.2
+* Fix dependencies
+
+# 0.1.1
+* Add travis integration
+* Extend detectjs to detect PhantomJS and SlimerJS
+* Add Bluebird for PhantomJS Promises
+* Add JQuery for PhantomJS
+* Simpler Test added (Just a div now)
+* Add SlimerJS launcher
+* Replace ES6 javascript for PhantomJS with plain javascript
+* Add Travis reporting tool
+
 # 0.1.0
 No CHANGELOG section found in Pull Request description.
 Use a `# CHANGELOG` section in your Pull Request description to auto-populate the `CHANGELOG.md`